Inventory items moved into physrisk; minor display improvements.
