+++ src/components/ChronicHazard.js
@@ -1,0 +1,18 @@
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import Typography from '@mui/material/Typography';
+import Title from './Title';
+
+function ChronicHazard(props) {
+    const { title, data, units } = props
+    return (
+        <React.Fragment>
+            <Title>{title}</Title>
+            <Typography color="text.secondary" sx={{ flex: 1, overflow: "auto" }}>
+                {data[0].y.toPrecision(6) + " " + units}
+            </Typography>
+        </React.Fragment>
+  );
+}
+
+export default ChronicHazard;
