+++ src/views/HazardViewer.js
@@ -1,22 +1,19 @@
 import { useEffect, useReducer, useRef, useState } from 'react';
 import Chart from '../components/Chart';
+import ChronicHazard from '../components/ChronicHazard';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import ScatterMap from '../components/ScatterMap';
 import Summary from '../components/Summary';
-import { hazardMenuReducer, loadHazardMenuData, updateMenuOptions } from '../data/HazardDataAvailability.js';
+import { hazardMenuReducer, hazardMenuInitialiser, loadHazardMenuData, updateMenuOptions } from '../data/HazardInventory.js';
 import {v4 as uuidv4} from 'uuid';
 import axios from 'axios';
 
 
 export default function HazardViewer(props) {
   const { visible } = props
-  const hazardMenuInitialState = {
-    inventory: null,
-    menus: null,
-    menuOptions: null,
-    selectedIndices: [0]
-  }
+  
+  const hazardMenuInitialState = hazardMenuInitialiser()
 
   // holds both the inventory of available hazard data and the menu state
   const [hazardMenu, hazardMenuUpdate] = useReducer(hazardMenuReducer, hazardMenuInitialState);
@@ -26,6 +23,7 @@ export default function HazardViewer(props) {
   const [lngLat, setLngLat] = useState(null);
   const apiHost = 'http://physrisk-api-sandbox.apps.odh-cl1.apps.os-climate.org';
 
+  //const apiHost = 'http://127.0.0.1:5000';
   function graphDataPoint(x, y) {
     return { x, y };
   }
@@ -46,25 +44,25 @@ export default function HazardViewer(props) {
     async function fetchGraphData() {
       if (hazardMenu.inventory)
       {
-        var [hazard, model, scenario, year] = hazardMenu.selection
         if (lngLat)
         {
           var payload = {
             "items": [
                 {
                     "request_item_id": uuidv4(),
-                    "event_type": hazard,
+                    "event_type": hazardMenu.selectedHazardTypeId,
                     "longitudes": [lngLat.lng],
                     "latitudes": [lngLat.lat],
-                    "year": year,
-                    "scenario": scenario.id,
-                    "model": model.id,
+                    "year": hazardMenu.selectedYear,
+                    "scenario": hazardMenu.selectedScenario.id,
+                    "model": hazardMenu.selectedModel.id,
                 },
             ],
           };
           var response = await axios.post(apiHost+'/api/get_hazard_data', payload);
           var curve_set = response.data.items[0].intensity_curve_set[0]
-          var points = curve_set.return_periods.map((item, i) => graphDataPoint(1.0/ item, curve_set.intensities[i]));
+          var points = (curve_set.intensities.length == 1) ? [graphDataPoint(0, curve_set.intensities[0])]
+            : curve_set.return_periods.map((item, i) => graphDataPoint(1.0/ item, curve_set.intensities[i]));
 
           setGraphData(points)
         }
@@ -72,6 +70,22 @@ export default function HazardViewer(props) {
     }
     fetchGraphData()
     }, [hazardMenu, lngLat]);
+
+  var chart
+  if (graphData)
+  {
+    var title = hazardMenu.menuOptions ? hazardMenu.menuOptions[1][hazardMenu.selectedIndices[1]] + 
+      (lngLat ? " @ (" + lngLat.lng.toFixed(4) + "\u00b0, " + lngLat.lat.toFixed(4) + "\u00b0)" : "") : ""
+    if (graphData.length > 1) 
+    {
+      chart = (<Chart title={title} data={graphData}/>)
+    }
+    else 
+    {
+      chart = (<ChronicHazard title={title} data={graphData} units={hazardMenu.selectedModel.units} />)
+    }
+  }
+  else chart = (<div></div>)
 
   return (
         <Grid container spacing={3}>
@@ -101,11 +115,7 @@ export default function HazardViewer(props) {
               height: 240,
             }}
           >
-            <Chart
-              title={hazardMenu.menuOptions ? hazardMenu.menuOptions[1][hazardMenu.selectedIndices[1]] + 
-                (lngLat ? " @ (" + lngLat.lng.toFixed(4) + "\u00b0, " + lngLat.lat.toFixed(4) + "\u00b0)" : "") : ""}
-              data={graphData}
-            />
+          {chart}
           </Paper>
         </Grid>
         {/* Summary */}
@@ -120,8 +130,7 @@ export default function HazardViewer(props) {
           >
             <Summary 
               modelName={hazardMenu.menuOptions ? hazardMenu.menuOptions[1][hazardMenu.selectedIndices[1]] : ""} 
-              modelDescription={data.current ? data.current.getModelDescription(hazardMenu.menuOptions[0][hazardMenu.selectedIndices[0]],
-                hazardMenu.menuOptions[1][hazardMenu.selectedIndices[1]]) : ""} 
+              modelDescription={hazardMenu?.selectedModel?.description} 
             />
           </Paper>
         </Grid>
